+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Automatic shooting when moving the player
+
+### Fixed
+- Fix issue where the pause menu wasn't properly working after a restart
 
 ## [1.0.1] - 2019-06-13
 ### Added